rustica::datatypes

Module async_monad

Async monad for asynchronous computations.

Asynchronous Monad

The AsyncM datatype represents an asynchronous computation that will eventually produce a value of type A. It provides a monadic interface for working with asynchronous operations in a functional programming style.

Functional Programming Context

In functional programming, asynchronous monads are used to:

Represent computations that will complete in the future

Compose and sequence asynchronous operations

Handle asynchronous control flow in a pure functional manner

Abstract away the complexity of async/await patterns

Similar constructs in other functional programming languages include:

IO in Cats Effect (Scala)

Task in Arrow (Kotlin)

Task in fp-ts (TypeScript)

IO in Haskell libraries like async

Type Class Implementations

The AsyncM type implements several important functional programming abstractions:

Functor: Allows mapping functions over the eventual result

Applicative: Enables applying functions wrapped in AsyncM to values wrapped in AsyncM

Monad: Provides sequencing of asynchronous operations

Basic Usage

use rustica::datatypes::async_monad::AsyncM;
use tokio;
use std::future::Future;

#[tokio::main]
async fn main() {
    // Create a pure value
    let async_value = AsyncM::pure(42);
     
    // Map over the value
    let doubled = async_value.clone().fmap(|x| async move { x * 2 });
    assert_eq!(doubled.try_get().await, 84);
     
    // Chain async computations
    let result = async_value
        .bind(|x| async move { AsyncM::pure(x + 1) })
        .fmap(|x| async move { x * 2 });
    assert_eq!(result.try_get().await, 86);
}

Structs

AsyncM

The asynchronous monad, which represents a computation that will eventually produce a value.

Type Aliases

BoxFuture

A type alias for an asynchronous computation that can be sent between threads.