rustica::datatypes::choice

Struct Choice

pub struct Choice<T> { /* private fields */ }

A type representing a value with multiple alternatives.

Choice<T> encapsulates a collection of values of type T. This structure is useful in scenarios where multiple options or choices need to be represented and manipulated.

Type Parameters

T: The type of the values stored in this Choice.

Fields

values: A vector containing all the values of type T.

Examples

use rustica::prelude::*;
use rustica::datatypes::choice::Choice;

let choice = Choice::new(1, vec![2, 3, 4]);
assert_eq!(*choice.first().unwrap(), 1);
assert_eq!(choice.alternatives(), &[2, 3, 4]);

Implementations

impl<T> Choice<T>

pub fn new_empty() -> Self

Creates a new empty Choice.

Examples

use rustica::datatypes::choice::Choice;

let empty: Choice<i32> = Choice::new_empty();
assert!(empty.is_empty());

pub fn new<I>(item: T, alternatives: I) -> Self
where I: IntoIterator<Item = T>, I::IntoIter: Iterator,

Creates a new Choice with a single item.

Examples

use rustica::datatypes::choice::Choice;

let choice = Choice::new(1, vec![2, 3, 4]);
assert_eq!(*choice.first().unwrap(), 1);
assert_eq!(choice.alternatives(), &[2, 3, 4]);

pub fn first(&self) -> Option<&T>

Returns a reference to the primary value.

Returns

A reference to the primary value of type T

Examples

use rustica::datatypes::choice::Choice;

let choice = Choice::new(42, vec![1, 2, 3]);
assert_eq!(*choice.first().unwrap(), 42);

pub fn alternatives(&self) -> &[T]

Get a reference to the alternatives (all items except the first).

Examples

use rustica::datatypes::choice::Choice;

let choice = Choice::new(1, vec![2, 3, 4]);
assert_eq!(choice.alternatives(), &[2, 3, 4]);

pub fn has_alternatives(&self) -> bool

Checks if there are any alternatives.

Returns

true if there are any alternatives, false otherwise.

Examples

use rustica::datatypes::choice::Choice;

let choice_with_alternatives = Choice::new(1, vec![2, 3]);
assert!(choice_with_alternatives.has_alternatives());

let choice_without_alternatives = Choice::new(1, vec![]);
assert!(!choice_without_alternatives.has_alternatives());

pub fn len(&self) -> usize

Returns the total number of choices, including the primary value and all alternatives.

Returns

The count of all choices as usize.

Examples

use rustica::datatypes::choice::Choice;

let choice = Choice::new(42, vec![1, 2, 3]);
assert_eq!(choice.len(), 4); // Primary value + 3 alternatives

pub fn is_empty(&self) -> bool

Checks if the Choice is empty (contains no values).

Returns

true if the Choice contains no values, false otherwise.

Examples

use rustica::datatypes::choice::Choice;

let empty_choice: Choice<i32> = Choice::new_empty();

assert!(empty_choice.is_empty());

let non_empty_choice = Choice::new(1, vec![2, 3]);
assert!(!non_empty_choice.is_empty());

pub fn add_alternatives<I>(self, items: I) -> Self
where T: Clone, I: IntoIterator<Item = T>,

Adds multiple new alternatives to the Choice, consuming the original.

This method creates a new Choice instance with the same primary value and all existing alternatives, plus the new items as additional alternatives.

Arguments

items - An iterator of values of type T to be added as new alternatives.

Returns

A new Choice<T> instance with the added alternatives.

Examples

use rustica::datatypes::choice::Choice;

let choice = Choice::new(1, vec![2, 3]);
let new_choice = choice.add_alternatives(vec![4, 5]);

assert_eq!(*new_choice.first().unwrap(), 1);
assert_eq!(new_choice.alternatives(), &[2, 3, 4, 5]);

pub fn remove_alternative(self, index: usize) -> Self
where T: Clone,

Removes an alternative at the specified index and returns a new Choice.

Arguments

index - The index of the alternative to remove (0-based, relative to the alternatives list).

Returns

A new Choice<T> with the alternative at the specified index removed.

Panics

Panics if the index is out of bounds.

Examples

use rustica::datatypes::choice::Choice;

let choice = Choice::new(1, vec![2, 3, 4]);
let new_choice = choice.remove_alternative(1);

assert_eq!(*new_choice.first().unwrap(), 1);
assert_eq!(new_choice.alternatives(), &[2, 4]);

pub fn filter<P>(&self, predicate: P) -> Self
where P: Fn(&T) -> bool, T: Clone,

Filters the alternatives based on a predicate.

Arguments

predicate - A function that takes a reference to an alternative and returns a boolean.

Returns

A new Choice with only the alternatives that satisfy the predicate.

Examples

use rustica::datatypes::choice::Choice;

let choice = Choice::new(1, vec![2, 3, 4, 5]);
let filtered = choice.filter(|x| x % 2 == 0);

assert_eq!(*filtered.first().unwrap(), 2);
assert_eq!(filtered.alternatives(), &[4]);

pub fn fmap_alternatives<F>(&self, f: F) -> Choice<T>
where F: Fn(&T) -> T, T: Clone,

Applies a function to all values in the Choice, including the primary value and alternatives.

Arguments

f - A function that takes a reference to T and returns a new T

Returns

A new Choice<T> with the function applied to all values

Examples

use rustica::datatypes::choice::Choice;

let choice = Choice::new(1, vec![2, 3, 4]);
let doubled = choice.fmap_alternatives(|&x| x * 2);

assert_eq!(*doubled.first().unwrap(), 1);
assert_eq!(doubled.alternatives(), &[4, 6, 8]);

pub fn flatten<I>(&self) -> Choice<I>
where T: IntoIterator<Item = I> + Clone, I: Clone,

Flattens a Choice of iterable items into a Choice of individual items.

This method transforms a Choice<T> where T is an iterable type into a Choice<T::Item>. The first item of the primary value becomes the new primary value, and all other items (including those from alternatives) are collected into the new alternatives.

Type Parameters

T: The original type, which must be clonable and iterable.

T::Item: The item type of the iterable, which must be clonable.

Returns

A new Choice<T::Item> with flattened contents.

Panics

Panics if the primary value is an empty iterator.

Examples

use rustica::datatypes::choice::Choice;

let nested = Choice::new(vec![1, 2], vec![vec![3, 4], vec![5]]);
let flattened = nested.flatten();

assert_eq!(*flattened.first().unwrap(), 1);
assert_eq!(flattened.alternatives(), &[3, 4, 5, 2]);

pub fn flatten_sorted<I>(&self) -> Choice<I>
where T: IntoIterator<Item = I> + Clone, I: Clone + Ord,

Flattens a Choice of iterable items into a sorted Choice of individual items.

Similar to flatten(), but sorts the alternatives according to their natural order. The first item of the primary value becomes the new primary value, and all other items (including those from alternatives) are sorted and collected into the new alternatives.

Type Parameters

T: The original type, which must be clonable and iterable.

I: The item type of the iterable, which must be clonable and implement Ord.

Returns

A new Choice<I> with flattened and sorted alternatives.

Panics

Panics if the primary value is an empty iterator.

Examples

use rustica::datatypes::choice::Choice;

let nested = Choice::new(vec![3, 1], vec![vec![4, 2], vec![5]]);
let flattened = nested.flatten_sorted();

assert_eq!(*flattened.first().unwrap(), 3);
assert_eq!(flattened.alternatives(), &[1, 2, 4, 5]);

pub fn of_many<I>(many: I) -> Self
where I: IntoIterator<Item = T>, T: Clone,

Creates a new Choice from a primary value and an iterable of alternatives.

Arguments

first - A value of type T to be used as the primary value

alternatives - An iterable containing items of type T to be used as alternatives

Returns

A new Choice<T> instance containing the primary value and alternatives

Examples

use rustica::datatypes::choice::Choice;

let choice = Choice::of_many(vec![1, 2, 3, 4]);
assert_eq!(*choice.first().unwrap(), 1);
assert_eq!(choice.alternatives(), &[2, 3, 4]);

pub fn filter_values<F>(&self, predicate: F) -> Self
where T: Clone, F: Fn(&T) -> bool,

Returns a Choice containing only the values that satisfy the predicate.

If the primary value doesn’t satisfy the predicate, the first alternative that does becomes the new primary value.

Arguments

predicate - A function that takes a reference to a value and returns a boolean

Returns

A new Choice with only the values that satisfy the predicate

Examples

use rustica::datatypes::choice::Choice;

let choice = Choice::new(1, vec![2, 3, 4, 5]);
let even = choice.filter_values(|x: &i32| x % 2 == 0);

assert_eq!(*even.first().unwrap(), 2);
assert_eq!(even.alternatives(), &[4]);

pub fn iter(&self) -> impl Iterator<Item = &T>

Returns an iterator over all values in the Choice, including the primary value and alternatives.

Returns

An iterator yielding references to all values of type &T.

Examples

use rustica::datatypes::choice::Choice;

let choice = Choice::new(1, vec![2, 3, 4]);
let values: Vec<&i32> = choice.iter().collect();
assert_eq!(values, vec![&1, &2, &3, &4]);

pub fn iter_alternatives(&self) -> impl Iterator<Item = &T>

Returns an iterator over the alternative values in the Choice, excluding the primary value.

Returns

An iterator yielding references to the alternative values of type &T.

Examples

use rustica::datatypes::choice::Choice;

let choice = Choice::new(1, vec![2, 3, 4]);
let alternatives: Vec<&i32> = choice.iter_alternatives().collect();
assert_eq!(alternatives, vec![&2, &3, &4]);

pub fn swap_with_alternative(self, alt_index: usize) -> Self
where T: Clone,

Swaps the first value with the alternative at the specified index, consuming the original.

Arguments

alt_index - The index of the alternative to swap with the first value

Returns

A new Choice with the first value swapped with the specified alternative

Examples

use rustica::datatypes::choice::Choice;

let choice = Choice::new(1, vec![2, 3, 4]);
let swapped = choice.swap_with_alternative(1);
assert_eq!(*swapped.first().unwrap(), 3);
assert_eq!(swapped.alternatives(), &[2, 1, 4]);