rustica::datatypes

Module iso_prism

Prism implementation for focusing on particular variants of a sum type.

Iso-based Prism

This module provides a Prism optic based on the Iso abstraction. A Prism is an optic that allows safe and functional access to a variant of a sum type (such as an enum), and the ability to construct the sum type from the focused value.

Core Idea

A Prism can be generalized as an Iso of the form S <-> Option<A>

preview/review functions are wrapped as Iso’s forward/backward operations

Examples

use rustica::datatypes::iso_prism::IsoPrism;
use rustica::traits::iso::Iso;

#[derive(Clone, Debug, PartialEq)]
enum MyEnum {
    Foo(i32),
    Bar(String),
}

struct FooPrismIso;
impl Iso<MyEnum, Option<i32>> for FooPrismIso {
    type From = MyEnum;
    type To = Option<i32>;
    fn forward(&self, from: &MyEnum) -> Option<i32> {
        match from {
            MyEnum::Foo(x) => Some(*x),
            _ => None,
        }
    }
    fn backward(&self, to: &Option<i32>) -> MyEnum {
        match to {
            Some(x) => MyEnum::Foo(*x),
            None => MyEnum::Bar("default".to_string()),
        }
    }
}

let prism = IsoPrism::new(FooPrismIso);
let foo = MyEnum::Foo(10);
assert_eq!(prism.preview(&foo), Some(10));
let bar = MyEnum::Bar("hi".to_string());
assert_eq!(prism.preview(&bar), None);
let reviewed = prism.review(&20);
assert_eq!(reviewed, MyEnum::Foo(20));

Lawful Optic Laws

Review-Preview Law: prism.preview(&prism.review(&a)) == Some(a)

Preview-Review Law: If prism.preview(s) == Some(a), then prism.review(&a) == s

See also: crate::datatypes::prism, crate::traits::iso::Iso

Structs

IsoPrism

Iso-based Prism optic.

LiftedPrismIso

Lifts a prism to work with Options.