rustica::datatypes::maybe

Enum Maybe

pub enum Maybe<T> {
    Just(T),
    Nothing,
}

A type that represents an optional value, optimized with null pointer optimization.

Maybe<T> is an enum that can be either Just(T) containing a value, or Nothing representing the absence of a value. It has the same memory layout as Option<T>.

Memory Optimization

This type uses Rust’s null pointer optimization. For types like Box<T>, Vec<T>, String, etc. that can never be null, the Nothing variant doesn’t require any additional memory. The compiler optimizes it to use the null pointer representation.

Type Parameters

T - The type of the value that might be contained

Examples

use rustica::datatypes::maybe::Maybe;

let just_five = Maybe::Just(5);
let nothing: Maybe<i32> = Maybe::Nothing;

// Pattern matching
match just_five {
    Maybe::Just(x) => println!("Got a value: {}", x),
    Maybe::Nothing => println!("Got nothing"),
}

Variants

Just(T)

Contains a value of type T

Nothing

Represents the absence of a value

Implementations

impl<T> Maybe<T>

pub const fn is_just(&self) -> bool

Returns true if the maybe value is a Just value.

Examples

use rustica::datatypes::maybe::Maybe;

let x: Maybe<u32> = Maybe::Just(2);
assert_eq!(x.is_just(), true);

let y: Maybe<u32> = Maybe::Nothing;
assert_eq!(y.is_just(), false);

pub const fn is_nothing(&self) -> bool

Returns true if the maybe value is a Nothing value.

Examples

use rustica::datatypes::maybe::Maybe;

let x: Maybe<u32> = Maybe::Just(2);
assert_eq!(x.is_nothing(), false);

let y: Maybe<u32> = Maybe::Nothing;
assert_eq!(y.is_nothing(), true);

pub fn from_option(opt: Option<T>) -> Self

Converts from Option<T> to Maybe<T>.

Some(x) is converted to Just(x), and None is converted to Nothing. This is a zero-cost conversion due to the same memory layout.

Examples

use rustica::datatypes::maybe::Maybe;

let opt_some = Some(42);
let opt_none: Option<i32> = None;

let maybe_just = Maybe::from_option(opt_some);
let maybe_nothing = Maybe::from_option(opt_none);

assert!(maybe_just.is_just());
assert!(maybe_nothing.is_nothing());

pub fn to_option(self) -> Option<T>

Converts from Maybe<T> to Option<T>.

Just(x) is converted to Some(x), and Nothing is converted to None. This is a zero-cost conversion due to the same memory layout.

Examples

use rustica::datatypes::maybe::Maybe;

let maybe_just = Maybe::Just(42);
let maybe_nothing: Maybe<i32> = Maybe::Nothing;

let opt_some = maybe_just.to_option();
let opt_none = maybe_nothing.to_option();

assert_eq!(opt_some, Some(42));
assert_eq!(opt_none, None);

pub fn to_result<E>(self, error: E) -> Result<T, E>

Converts this Maybe to a Result with the specified error.

Just(x) is converted to Ok(x).

Nothing is converted to Err(error).

Arguments

error - The error to use if this is Nothing

Examples

use rustica::datatypes::maybe::Maybe;

let maybe_just = Maybe::Just(42);
let maybe_nothing: Maybe<i32> = Maybe::Nothing;

let ok_result = maybe_just.to_result("No value present");
let err_result = maybe_nothing.to_result("No value present");

assert_eq!(ok_result, Ok(42));
assert_eq!(err_result, Err("No value present"));

pub fn to_standard_result(self) -> Result<T, MaybeError>

Converts this Maybe to a Result with a standard MaybeError.

Just(x) is converted to Ok(x).

Nothing is converted to Err(MaybeError::ValueNotPresent).

Examples

use rustica::datatypes::maybe::{Maybe, MaybeError};

let maybe_just = Maybe::Just(42);
let maybe_nothing: Maybe<i32> = Maybe::Nothing;

let ok_result = maybe_just.to_standard_result();
let err_result = maybe_nothing.to_standard_result();

assert_eq!(ok_result.is_ok(), true);
assert_eq!(err_result.unwrap_err(), MaybeError::ValueNotPresent);

pub fn try_unwrap(self) -> Result<T, AppError<&'static str, &'static str>>

Unwraps a maybe, yielding the content of a Just.

Errors

If the value is a Nothing, this function generates and returns an error with a detailed message.

Examples

use rustica::datatypes::maybe::Maybe;
use rustica::utils::error_utils::ResultExt;

let x = Maybe::Just("air");
let result = x.try_unwrap().unwrap_or_default();
assert_eq!(result, "air");

let y: Maybe<&str> = Maybe::Nothing;
let result = y.try_unwrap();
assert!(result.is_err());

pub fn unwrap(self) -> T

Unwraps a maybe, yielding the content of a Just.

Panics

Panics if the value is a Nothing with a custom panic message.

Examples

use rustica::datatypes::maybe::Maybe;

let x = Maybe::Just("air");
assert_eq!(x.unwrap(), "air");

use rustica::datatypes::maybe::Maybe;

let x: Maybe<&str> = Maybe::Nothing;
x.unwrap(); // This will panic

pub fn unwrap_or(self, default: T) -> T

Unwraps a maybe, yielding the content of a Just or a default value.

Examples

use rustica::datatypes::maybe::Maybe;

let just = Maybe::Just(42);
let nothing: Maybe<i32> = Maybe::Nothing;

assert_eq!(just.unwrap_or(0), 42);
assert_eq!(nothing.unwrap_or(0), 0);

pub fn fmap_or<U, F>(self, default: U, f: F) -> U
where F: FnOnce(T) -> U,

Returns the provided default value if Nothing, or applies a function to the contained value.

Examples

use rustica::datatypes::maybe::Maybe;

let maybe_some = Maybe::Just(41);
let maybe_none: Maybe<i32> = Maybe::Nothing;

let incremented = maybe_some.fmap_or(0, |x| x + 1);
let default = maybe_none.fmap_or(0, |x| x + 1);

assert_eq!(incremented, 42);
assert_eq!(default, 0);

pub const fn as_ref(&self) -> Option<&T>

Converts a Maybe<T> to an Option<&T>.

Converts Just(x) to Some(&x) and Nothing to None.

Examples

use rustica::datatypes::maybe::Maybe;

let maybe_just = Maybe::Just(42);
let maybe_nothing: Maybe<i32> = Maybe::Nothing;

assert_eq!(maybe_just.as_ref(), Some(&42));
assert_eq!(maybe_nothing.as_ref(), None);