rustica::datatypes::wrapper

Module last

Last

This module provides the Last wrapper type which forms a semigroup by taking the last non-None value.

use rustica::datatypes::wrapper::last::Last;
use rustica::traits::semigroup::Semigroup;

let a = Last(Some(5));
let b = Last(Some(10));
let c = a.combine(&b);
assert_eq!(c, Last(Some(10))); // Takes the last value

let x = Last(None);
let y = Last(Some(7));
let z = x.combine(&y);
assert_eq!(z, Last(Some(7))); // First value was None, so takes the second

Structs

Last

A wrapper type that forms a semigroup by taking the last non-None value.