rustica::datatypes::wrapper::last

Struct Last

#[repr(transparent)]
pub struct Last<T>(pub Option<T>);

A wrapper type that forms a semigroup by taking the last non-None value.

The monoid instance uses None as the identity element.

Examples

use rustica::datatypes::wrapper::last::Last;
use rustica::traits::semigroup::Semigroup;
use rustica::traits::monoid::Monoid;

let a = Last(Some(5));
let b = Last(Some(7));
let c = a.combine(&b);
assert_eq!(c, Last(Some(7)));

// Last is associative
let x = Last(Some(1));
let y = Last(None);
let z = Last(Some(3));
assert_eq!(x.clone().combine(&y.clone()).combine(&z.clone()),
           x.clone().combine(&y.clone()).combine(&z.clone()));

// Identity element
let id = Last::empty();  // Last(None)
assert_eq!(id, Last(None));
assert_eq!(Last(Some(42)).combine(&id.clone()), Last(Some(42)));
assert_eq!(id.combine(&Last(Some(42))), Last(Some(42)));

Tuple Fields
0: Option<T>