rustica::datatypes::wrapper

Module max

Max

This module provides the Max wrapper type which forms a semigroup under taking the maximum.

use rustica::datatypes::wrapper::max::Max;
use rustica::traits::semigroup::Semigroup;

let a = Max(5);
let b = Max(10);
let c = a.combine(&b);
assert_eq!(c, Max(10));

Structs

Max

A wrapper type that forms a semigroup under the maximum operation.