rustica::datatypes::wrapper

Module memoize

Memoize

This module provides the Memoize wrapper type that caches the result of a function call.

Structs

MemoizeDeprecated

A wrapper that memoizes (caches) the result of a function.

MemoizeFnDeprecated

A wrapper that memoizes (caches) the result of a function that takes an input.

ThreadSafeMemoizeDeprecated

A thread-safe version of Memoize that can be safely shared across threads.

ThreadSafeMemoizeFnDeprecated

A wrapper that memoizes (caches) the result of a function that takes an input.