rustica::datatypes::wrapper

Module thunk

A lightweight thunk that can be evaluated.

This module provides the Thunk type, which is a statically-typed function wrapper that implements the Evaluate trait.

Examples

use rustica::traits::evaluate::{Evaluate, EvaluateExt};
use rustica::datatypes::wrapper::thunk::Thunk;

// Create a thunk that produces a value
let computation = Thunk::new(|| 42);

// Evaluate by reference
assert_eq!(computation.evaluate(), 42);

// Using extension methods
let string_result: String = computation.fmap_evaluate(|x| x.to_string());
assert_eq!(string_result, "42");

// Evaluate by consuming the thunk
let result: i32 = computation.evaluate_owned();
assert_eq!(result, 42);

Structs

Thunk

A thunk that lazily produces a value when evaluated.