rustica::datatypes::wrapper::value

Struct Value

#[repr(transparent)]
pub struct Value<T>(pub T);

A simple value container that just returns its wrapped value when evaluated.

This type is useful for:

Creating constant values that adhere to the Evaluate interface

Converting regular values to work with evaluation pipelines

Unifying different evaluation sources in a common interface

Type Parameters

T - The type of the contained value

Tuple Fields
0: T
Implementations

impl<T> Value<T>

pub fn new(value: T) -> Self

Creates a new Value wrapper.

Parameters

value - The value to wrap

Returns

A new Value instance containing the given value