rustica::traits::composable

Function compose_async

pub async fn compose_async<A, B, C, F, G, FutB, FutC>(f: F, g: G, input: A) -> C
where
    F: Fn(A) -> FutB,
    G: Fn(B) -> FutC,
    FutB: Future<Output = B>,
    FutC: Future<Output = C>,

Composes two asynchronous functions to create a new asynchronous function.

This function applies two asynchronous functions in sequence, first applying f to the input and then applying g to the result.

Type Parameters

A: Input type of the first function

B: Output type of the first function and input type of the second function

C: Output type of the second function

F: Function type that takes A and returns Future<Output = B>

G: Function type that takes B and returns Future<Output = C>

FutB: Future type returned by F

FutC: Future type returned by G

Arguments

f: First asynchronous function

g: Second asynchronous function

input: Initial input value

Returns

The result C after sequentially applying f and g

Examples

use rustica::traits::composable::compose_async;

async fn add_one(x: i32) -> i32 { x + 1 }
async fn double(x: i32) -> i32 { x * 2 }

let result = compose_async(add_one, double, 3).await;
assert_eq!(result, 8);  // (3 + 1) * 2 = 8