rustica::traits::composable

Function compose_fallible

pub fn compose_fallible<A, B, C, E, F, G>(
    f: F,
    g: G,
) -> impl Fn(A) -> Result<C, E>
where
    F: Fn(A) -> Result<B, E>,
    G: Fn(B) -> Result<C, E>,

Composes two fallible functions that return Results.

This function composes two functions where both can return errors. If the first function returns an error, that error is propagated. Otherwise, the second function is applied to the success value of the first function.

Type Parameters

A: Input type of the first function

B: Intermediate type (output of first function, input to second function)

C: Output type of the second function

E: Error type for both functions

F: Type of the first function

G: Type of the second function

Arguments

f: First fallible function

g: Second fallible function

Returns

A new function that returns a Result, propagating errors or applying both functions

Examples

use rustica::traits::composable::compose_fallible;
use std::num::ParseIntError;

// Define two fallible functions
fn parse_string(s: &str) -> Result<i32, ParseIntError> {
    s.parse::<i32>()
}

fn double_if_positive(n: i32) -> Result<i32, &'static str> {
    if n > 0 {
        Ok(n * 2)
    } else {
        Err("Number was not positive")
    }
}

// Compose them with error type conversion
let composed = compose_fallible(
    |s| parse_string(s).map_err(|_| "Parse error"),
    double_if_positive
);

assert_eq!(composed("5"), Ok(10));
assert_eq!(composed("-5"), Err("Number was not positive"));
assert_eq!(composed("not_a_number"), Err("Parse error"));