rustica::traits::composable

Function compose_iter_chain

pub fn compose_iter_chain<A, B, F>(functions: Vec<F>) -> impl Fn(A) -> Vec<B>
where
    F: Fn(A) -> Vec<B> + Clone,
    A: Clone,

Chains multiple iterator-producing functions into a single function.

Type Parameters

A: Input type to all functions

B: Output item type of all functions

F: Type of the functions

Arguments

functions: Vector of functions that each produce a collection

Returns

A function that chains all the iterators produced by the given functions

Examples

use rustica::traits::composable::compose_iter_chain;

// Define iterator-producing functions for different ranges
fn range_1_to_3(_: ()) -> Vec<i32> { vec![1, 2, 3] }
fn range_4_to_6(_: ()) -> Vec<i32> { vec![4, 5, 6] }
fn range_7_to_9(_: ()) -> Vec<i32> { vec![7, 8, 9] }

// Chain them together
let combined = compose_iter_chain(vec![range_1_to_3, range_4_to_6, range_7_to_9]);

// Use the combined function
let all_numbers: Vec<_> = combined(());

// Contains all numbers from all three iterators, in order of the iterators
// First all numbers from range1 (1..100), then all numbers from range2 (100..200), then all numbers from range3 (200..300)
assert_eq!(all_numbers.len(), 9);