rustica::traits::composable

Function compose_option

pub fn compose_option<A, B, C, F, G>(f: F, g: G) -> impl Fn(A) -> Option<C>
where
    F: Fn(A) -> Option<B>,
    G: Fn(B) -> Option<C>,

Composes two functions that return Options.

This function is similar to using and_then on an Option. If the first function returns None, that None is propagated. Otherwise, the second function is applied to the value inside the Some.

Type Parameters

A: Input type of the first function

B: Intermediate type (output of first function, input to second function)

C: Output type of the second function

F: Type of the first function

G: Type of the second function

Arguments

f: First function returning Option

g: Second function returning Option

Returns

A new function that returns an Option, propagating None or applying both functions

Examples

use rustica::traits::composable::compose_option;

// Define two functions that return Options
fn parse_as_int(s: &str) -> Option<i32> {
    s.parse::<i32>().ok()
}

fn double_if_positive(n: i32) -> Option<i32> {
    if n > 0 {
        Some(n * 2)
    } else {
        None
    }
}

// Compose them
let composed = compose_option(parse_as_int, double_if_positive);

assert_eq!(composed("5"), Some(10));
assert_eq!(composed("-5"), None);         // Not positive
assert_eq!(composed("not_a_number"), None); // Parse error