rustica::traits::contravariant_functor

Trait ContravariantFunctor

pub trait ContravariantFunctor: HKT {
    // Required methods
    fn contramap<U, F>(&self, f: F) -> Self::Output<U>
       where F: Fn(&U) -> Self::Source,
             U: Clone;
    fn contramap_owned<U, F>(self, f: F) -> Self::Output<U>
       where F: Fn(U) -> Self::Source,
             U: Clone,
             Self: Sized;
}

A contravariant functor is a type constructor that allows mapping functions in a way that reverses their direction. While regular functors map functions forward (A -> B), contravariant functors map functions backward (B -> A).

Required Methods

fn contramap<U, F>(&self, f: F) -> Self::Output<U>
where F: Fn(&U) -> Self::Source, U: Clone,

Maps a function that transforms values of type U into values of type Self::Source, producing a new contravariant functor that works with type U.

This is the reference-based version of contramap, which only takes a reference to self.

Type Parameters

U: The new input type for the resulting functor

Arguments

f: Function that converts from references of the new type U to Self::Source

Returns

A new contravariant functor that works with values of type U

fn contramap_owned<U, F>(self, f: F) -> Self::Output<U>
where F: Fn(U) -> Self::Source, U: Clone, Self: Sized,

Maps a function that transforms values of type U into values of type Self::Source, producing a new contravariant functor that works with type U.

This is the ownership-based version of contramap, which takes ownership of self.

Type Parameters

U: The new input type for the resulting functor

Arguments

f: Function that converts from the new type U to Self::Source

Returns

A new contravariant functor that works with values of type U

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.