rustica::traits::contravariant_functor

Trait ContravariantFunctorExt

pub trait ContravariantFunctorExt: ContravariantFunctor {
    // Provided methods
    fn contramap_with<U, F>(&self, f: F) -> Self::Output<U>
       where F: Fn(&U) -> Self::Source,
             U: Clone { ... }
    fn contramap_const<U>(&self, value: Self::Source) -> Self::Output<U>
       where Self::Source: Clone,
             U: Clone { ... }
    fn contra_compose<G>(&self, _g: &G) -> Self::Output<G::Source>
       where G: ContravariantFunctor,
             G::Source: Clone,
             Self::Source: From<G::Source> { ... }
}

Extension methods for contravariant functors, providing additional utility functions.

Provided Methods

fn contramap_with<U, F>(&self, f: F) -> Self::Output<U>
where F: Fn(&U) -> Self::Source, U: Clone,

Chains contramap operations for convenient method chaining.

This method makes it easier to chain multiple contramap operations together.

Type Parameters

U: The new input type for the resulting functor

Arguments

f: Function that converts from references of the new type U to Self::Source

Returns

A new contravariant functor that works with values of type U

fn contramap_const<U>(&self, value: Self::Source) -> Self::Output<U>
where Self::Source: Clone, U: Clone,

Contramaps with a function that ignores its input, effectively creating a constant function.

This is useful when you want to create a contravariant functor that always uses a specific value regardless of the input.

Type Parameters

U: The new input type for the resulting functor

Arguments

value: The constant value to use for all inputs

Returns

A new contravariant functor that works with values of type U

fn contra_compose<G>(&self, _g: &G) -> Self::Output<G::Source>
where G: ContravariantFunctor, G::Source: Clone, Self::Source: From<G::Source>,

Composes two contravariant functors, creating a chain of transformations.

Type Parameters

G: Another contravariant functor

Arguments

g: Another contravariant functor

Returns

A composed contravariant functor

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.