rustica::traits

Module hkt

Higher-kinded type abstractions for generic programming. The HKT trait enables emulation of higher-kinded types in Rust.

Higher-kinded types (HKTs) are type constructors that take a type and return another type. Rust does not directly support higher-kinded types, but we can emulate them using associated types.

This module provides the HKT trait and related traits that form the foundation for higher-kinded polymorphism in the Rustica library.

Examples

use rustica::traits::hkt::HKT;

// Define our own wrapper types for the examples
// MyOption is a simple wrapper around Option
#[derive(Clone)]
struct MyOption<T>(Option<T>);

// Implement HKT for our custom wrapper
impl<T> HKT for MyOption<T> {
    type Source = T;
    type Output<U> = MyOption<U>;
}

// MyVec is a simple wrapper around Vec
#[derive(Clone)]
struct MyVec<T>(Vec<T>);

// Implement HKT for our custom wrapper
impl<T> HKT for MyVec<T> {
    type Source = T;
    type Output<U> = MyVec<U>;
}

// Writing a function that works with any HKT
fn map_hkt<H, T, U, F>(value: &H, f: F) -> H::Output<U>
where
    H: HKT<Source = T>,
    F: Fn(&T) -> U,
{
    // This would be the implementation in a real case
    // Here we're just demonstrating the type signatures
    unimplemented!()
}

Structs

HKTType

A phantom type used to represent a higher-kinded type at the type level.

Traits

BinaryHKT

A trait for higher-kinded types that have two type parameters.

HKT

A trait for types that can be treated as higher-kinded types.