rustica::traits::hkt

Trait BinaryHKT

pub trait BinaryHKT: HKT {
    type Source2;
    type BinaryOutput<Type1, Type2>: BinaryHKT<Source = Type1, Source2 = Type2>;

    // Required methods
    fn map_second<F, NewType2>(
        &self,
        f: F,
    ) -> Self::BinaryOutput<Self::Source, NewType2>
       where F: Fn(&Self::Source2) -> NewType2,
             Self::Source: Clone,
             Self::Source2: Clone,
             NewType2: Clone;
    fn map_second_owned<F, NewType2>(
        self,
        f: F,
    ) -> Self::BinaryOutput<Self::Source, NewType2>
       where F: Fn(Self::Source2) -> NewType2,
             NewType2: Clone;
}

A trait for higher-kinded types that have two type parameters.

This trait extends the HKT trait to allow for types that have a second type parameter, such as Result<T, E> or Either<L, R>.

Examples

use rustica::traits::hkt::{HKT, BinaryHKT};

// Define our own wrapper types for the examples
// MyResult is a simple wrapper around Result
#[derive(Clone)]
struct MyResult<T, E>(Result<T, E>);

// Implement HKT for our custom wrapper
impl<T, E> HKT for MyResult<T, E> {
    type Source = T;
    type Output<U> = MyResult<U, E>;
}

// Implement BinaryHKT for our custom wrapper
impl<T, E> BinaryHKT for MyResult<T, E> {
    type Source2 = E;
    type BinaryOutput<Type1, Type2> = MyResult<Type1, Type2>;
     
    fn map_second<F, NewType2>(&self, f: F) -> Self::BinaryOutput<Self::Source, NewType2>
    where
        F: Fn(&Self::Source2) -> NewType2,
        Self::Source: Clone,
        Self::Source2: Clone,
    {
        MyResult(match &self.0 {
            Ok(v) => Ok(v.clone()),
            Err(e) => Err(f(e)),
        })
    }

    fn map_second_owned<F, NewType2>(self, f: F) -> Self::BinaryOutput<Self::Source, NewType2>
    where
        F: Fn(Self::Source2) -> NewType2,
    {
        MyResult(match self.0 {
            Ok(v) => Ok(v),
            Err(e) => Err(f(e)),
        })
    }
}

// A function that works with BinaryHKT
fn map_second_generic<H, T, E, U, F>(value: &H, f: F) -> H::BinaryOutput<T, U>
where
    H: BinaryHKT<Source = T, Source2 = E>,
    F: Fn(&E) -> U,
    T: Clone,
    E: Clone,
    U: Clone,
{
    value.map_second(f)
}

Required Associated Types

type Source2

The second type parameter of this HKT.

type BinaryOutput<Type1, Type2>: BinaryHKT<Source = Type1, Source2 = Type2>

The same HKT but with both type parameters replaced.

Required Methods

fn map_second<F, NewType2>( &self, f: F, ) -> Self::BinaryOutput<Self::Source, NewType2>
where F: Fn(&Self::Source2) -> NewType2, Self::Source: Clone, Self::Source2: Clone, NewType2: Clone,

Maps a function over the second type parameter.

Type Parameters

F - The function type

NewType2 - The type of the transformed second parameter

Parameters

f - A function that transforms the second type parameter

Returns

A new HKT with the second type parameter transformed

fn map_second_owned<F, NewType2>( self, f: F, ) -> Self::BinaryOutput<Self::Source, NewType2>
where F: Fn(Self::Source2) -> NewType2, NewType2: Clone,

Maps a function over the second type parameter, consuming the HKT.

Type Parameters

F - The function type

NewType2 - The type of the transformed second parameter

Parameters

f - A function that transforms the second type parameter

Returns

A new HKT with the second type parameter transformed

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.