rustica::traits::hkt

Trait HKT

pub trait HKT {
    type Source;
    type Output<NewType>: HKT<Source = NewType>;
}

A trait for types that can be treated as higher-kinded types.

In category theory, a functor is a mapping between categories. In Rust terms, it can be seen as a container type that can be transformed while preserving its structure.

The HKT trait provides a way to refer to the contained type and to construct the same container with a different contained type.

Type Parameters

Source - The type contained in this HKT

Output<U> - The same HKT but containing type U instead of Source

Examples

use rustica::traits::hkt::HKT;

// Define our own wrapper types for the examples
// MyOption is a simple wrapper around Option
#[derive(Clone)]
struct MyOption<T>(Option<T>);

// Implement HKT for our custom wrapper
impl<T> HKT for MyOption<T> {
    type Source = T;
    type Output<U> = MyOption<U>;
}

// MyVec is a simple wrapper around Vec
#[derive(Clone)]
struct MyVec<T>(Vec<T>);

// Implement HKT for our custom wrapper
impl<T> HKT for MyVec<T> {
    type Source = T;
    type Output<U> = MyVec<U>;
}

// A function that uses HKT
fn transform<H, T, U>(container: &H, value: &T) -> H::Output<U>
where
    H: HKT<Source = T>,
    T: Clone,
    U: From<T>,
{
    // Just an example signature
    unimplemented!()
}

Required Associated Types

type Source

The type contained in this HKT.

type Output<NewType>: HKT<Source = NewType>

The same HKT but containing type NewType instead of Source.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.