rustica::traits::identity

Trait Identity

pub trait Identity: HKT {
    // Required methods
    fn value(&self) -> &Self::Source;
    fn into_value(self) -> Self::Source;
    fn pure_identity<A>(value: A) -> Self::Output<A>
       where Self::Output<A>: Identity,
             A: Clone;

    // Provided methods
    fn try_value(&self) -> Option<&Self::Source> { ... }
    fn try_into_value(self) -> Option<Self::Source>
       where Self: Sized { ... }
    fn id<A>(a: A) -> A { ... }
}

A trait for types that represent identity functions in category theory.

In category theory, an identity morphism (or identity function) is a morphism that leaves an object unchanged. The Identity trait provides functionality for working with identity functions and accessing values in a type-safe way.

Laws

Types implementing Identity should adhere to the following laws:

Left identity: identity.pure_identity(a).map(f) == f(a)

Right identity: identity.map(Identity::id) == identity

Examples

use rustica::traits::hkt::HKT;
use rustica::traits::identity::Identity;

// A simple wrapper type
struct Wrapper<T>(T);

impl<T> HKT for Wrapper<T> {
    type Source = T;
    type Output<U> = Wrapper<U>;
}

impl<T> Identity for Wrapper<T> {
    fn value(&self) -> &Self::Source {
        &self.0
    }
     
    fn into_value(self) -> Self::Source {
        self.0
    }

    fn pure_identity<A>(value: A) -> Self::Output<A> {
        Wrapper(value)
    }
}

// Using the Identity trait
let wrapped: Wrapper<i32> = Wrapper(42);
assert_eq!(*wrapped.value(), 42);

// Using the identity function
let x: i32 = 5;
assert_eq!(<Wrapper<i32> as Identity>::id(x), 5);

Required Methods

fn value(&self) -> &Self::Source

Returns a reference to the contained value.

Panics

This method may panic if the container doesn’t contain a valid value. For a non-panicking alternative, use try_value().

Examples

use rustica::traits::identity::Identity;

// For an Option:
let some_value: Option<i32> = Some(42);
assert_eq!(*some_value.value(), 42);

// This would panic:
// let none_value: Option<i32> = None;
// none_value.value(); // panics

fn into_value(self) -> Self::Source

Consumes self and returns the contained value.

Panics

This method may panic if the container doesn’t contain a valid value. For a non-panicking alternative, use try_into_value().

Examples

use rustica::traits::identity::Identity;

// For an Option:
let some_value: Option<i32> = Some(42);
assert_eq!(some_value.into_value(), 42);

// This would panic:
// let none_value: Option<i32> = None;
// none_value.into_value(); // panics

fn pure_identity<A>(value: A) -> Self::Output<A>
where Self::Output<A>: Identity, A: Clone,

Creates an identity instance containing the given value.

This is a convenience method for creating a new instance of a type that implements Identity.

Type Parameters

A: The type of the value to wrap

Arguments

value: The value to wrap

Examples

use rustica::traits::identity::Identity;

// For an Option:
let value: i32 = 42;
let option = <Option<i32> as Identity>::pure_identity(value);
assert_eq!(option, Some(42));

Provided Methods

fn try_value(&self) -> Option<&Self::Source>

Returns a reference to the contained value, if available.

This method is safer than value() as it returns an Option instead of potentially panicking.

Default Implementation

By default, this calls value() and wraps the result in Some. Types that might not contain a value should override this method.

Examples

use rustica::traits::identity::Identity;

// For an Option:
let some_value: Option<i32> = Some(42);
assert_eq!(some_value.try_value(), Some(&42));

let none_value: Option<i32> = None;
assert_eq!(none_value.try_value(), None);

fn try_into_value(self) -> Option<Self::Source>
where Self: Sized,

Consumes self and returns the contained value, if available.

This method is safer than into_value() as it returns an Option instead of potentially panicking.

Default Implementation

By default, this method attempts to obtain the value using into_value(). Types that might not contain a value should override this method.

Examples

use rustica::traits::identity::Identity;

// For an Option:
let some_value: Option<i32> = Some(42);
assert_eq!(some_value.try_into_value(), Some(42));

let none_value: Option<i32> = None;
assert_eq!(none_value.try_into_value(), None);

fn id<A>(a: A) -> A

The identity function, which returns its input unchanged.

This function serves as the identity morphism in category theory.

Type Parameters

A: The type of the input value

Arguments

a: The value to return unchanged

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.