rustica::traits::iso

Trait IsoExt

pub trait IsoExt<A, B>: Iso<A, B> {
    // Provided methods
    fn convert_forward(&self, value: &Self::From) -> Self::To { ... }
    fn convert_backward(&self, value: &Self::To) -> Self::From { ... }
    fn modify<F>(&self, from: &Self::From, f: F) -> Self::From
       where F: FnOnce(Self::To) -> Self::To { ... }
    fn verify_laws(&self, from: &Self::From, to: &Self::To) -> bool
       where Self::From: PartialEq,
             Self::To: PartialEq { ... }
}

Extension methods for types that implement Iso.

Provided Methods

fn convert_forward(&self, value: &Self::From) -> Self::To

Applies this isomorphism to convert a value of the source type into the target type.

Arguments

value - A value of the source type

Returns

The converted value in the target type

fn convert_backward(&self, value: &Self::To) -> Self::From

Applies this isomorphism to convert a value of the target type back to the source type.

Arguments

value - A value of the target type

Returns

The converted value in the source type

fn modify<F>(&self, from: &Self::From, f: F) -> Self::From
where F: FnOnce(Self::To) -> Self::To,

Modifies a value of the source type by applying a function to its representation in the target type.

Type Parameters

F - The type of the function that modifies target values

Arguments

from - A value of the source type

f - A function that transforms a target value into another target value

Returns

The modified value in the source type

fn verify_laws(&self, from: &Self::From, to: &Self::To) -> bool
where Self::From: PartialEq, Self::To: PartialEq,

Verifies that this isomorphism satisfies the isomorphism laws for the given values.

Arguments

from - A value of the source type

to - A value of the target type

Returns

true if the isomorphism laws are satisfied for the given values, false otherwise

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.