rustica::traits

Module natural_transformation

Transformations between functors.

Natural Transformation Trait

A natural transformation is a morphism between two functors, preserving structure across all types.

This trait enables generic, type-safe transformations between higher-kinded types (functors).

Example

use rustica::datatypes::maybe::Maybe;
use rustica::datatypes::either::Either;
use rustica::traits::natural_transformation::NaturalTransformation;

struct MaybeToEither;
impl<L: Clone + Default, A: Clone> NaturalTransformation<Maybe<A>, Either<L, A>> for MaybeToEither {
    fn transform<T: Clone>(&self, fa: &Maybe<T>) -> Either<L, T> {
        match fa {
            Maybe::Just(a) => Either::Right(a.clone()),
            Maybe::Nothing => Either::Left(L::default()),
        }
    }
}

Traits

NaturalTransformation

A natural transformation between two functors F and G.

Functions

identity_nat

Returns the identity natural transformation for any functor.