rustica::traits::semigroup

Function combine_values

pub fn combine_values<T, I>(initial: T, values: I) -> T
where
    T: Semigroup,
    I: IntoIterator<Item = T>,

Combines a sequence of semigroup values, starting with an initial value.

Type Parameters

T - A type that implements Semigroup

I - An iterator type that yields values of type T

Parameters

initial - The initial value to start combining with

values - An iterator of values to combine with the initial value

Returns

The result of combining the initial value with all values in the iterator

Examples

use rustica::traits::semigroup::{self, Semigroup};
use rustica::datatypes::wrapper::product::Product;

let initial = Product(2);
let values = vec![Product(3), Product(4)];
let result = semigroup::combine_values(initial, values);
assert_eq!(result, Product(24));  // 2 * 3 * 4 = 24