rustica::traits

Module traversable

Traversing data structures with effects.

Traversable

A trait for structures that can be traversed from left to right while preserving structure and accumulating effects in an applicative functor.

Mathematical Definition

For a type constructor T, a traversable instance consists of two operations:

traverse: (A -> F<B>) -> T<A> -> F<T<B>>

sequence: T<F<A>> -> F<T<A>>

where F is an applicative functor.

TODO

Add benchmarks for traversal operations

Implement the Traversable trait for common types (Option, Result, Vec)

Implement TraversableExt with additional utility methods such as:

for_each: Apply function for side effects only

flat_map: Map and flatten in one operation

traverse_map: Combine traverse and map

filter_a: Filter with applicative effects

zip_with: Combine two traversables

Traits

Traversable

A trait for structures that can be traversed from left to right while preserving structure and accumulating effects in an applicative functor.