rustica::traits::traversable

Trait Traversable

pub trait Traversable: Applicative {
    // Required methods
    fn traverse<F, B, Func>(&self, f: Func) -> F::Output<Self::Output<B>>
       where F: Applicative,
             Func: Fn(&Self::Source) -> F::Output<B>;
    fn traverse_owned<F, B, Func>(self, f: Func) -> F::Output<Self::Output<B>>
       where F: Applicative,
             Func: FnOnce(Self::Source) -> F::Output<B>,
             Self: Sized;

    // Provided methods
    fn sequence<F>(&self) -> F::Output<Self::Output<F::Source>>
       where F: Applicative,
             Self::Source: Into<F::Output<F::Source>> + Clone { ... }
    fn sequence_owned<F>(self) -> F::Output<Self::Output<F::Source>>
       where F: Applicative,
             Self::Source: Into<F::Output<F::Source>>,
             Self: Sized { ... }
}

A trait for structures that can be traversed from left to right while preserving structure and accumulating effects in an applicative functor.

Type Parameters

Source is the type of elements in the traversable structure

Output<T> represents the structure containing elements of type T

Required Methods

fn traverse<F, B, Func>(&self, f: Func) -> F::Output<Self::Output<B>>
where F: Applicative, Func: Fn(&Self::Source) -> F::Output<B>,

Traverses the structure, applying the given function to each element and collecting the results.

This method allows you to apply an effect-producing function to each element in the structure, and then collect all of these effects into a single effect containing the new structure.

Type Parameters

F: The applicative functor representing the effect

B: The type of elements in the resulting structure

Func: The type of the function to apply to each element

Arguments

f: A function that takes a reference to an element of type Self::Source and returns an effect F::Output<B>

Returns

An effect F::Output containing the new structure Self::Output<B>

fn traverse_owned<F, B, Func>(self, f: Func) -> F::Output<Self::Output<B>>
where F: Applicative, Func: FnOnce(Self::Source) -> F::Output<B>, Self: Sized,

Traverses the structure with ownership, applying the given function to each element and collecting the results.

Similar to traverse, but takes ownership of the traversable structure, which can avoid unnecessary clones in certain situations.

Type Parameters

F: The applicative functor representing the effect

B: The type of elements in the resulting structure

Func: The type of the function to apply to each element

Arguments

f: A function that takes an element of type Self::Source and returns an effect F::Output<B>

Returns

An effect F::Output containing the new structure Self::Output<B>

Provided Methods

fn sequence<F>(&self) -> F::Output<Self::Output<F::Source>>
where F: Applicative, Self::Source: Into<F::Output<F::Source>> + Clone,

Sequences a structure of effects into an effect of structure.

This operation takes a structure where each element is an effect and reorders it into a single effect containing a structure of values. It is equivalent to traverse(identity) when the elements are already effects.

Type Parameters

F: The applicative functor containing the effects

Returns

The reordered structure wrapped in a single effect

fn sequence_owned<F>(self) -> F::Output<Self::Output<F::Source>>
where F: Applicative, Self::Source: Into<F::Output<F::Source>>, Self: Sized,

Sequences a structure of effects with ownership, into an effect of structure.

Similar to sequence_ref, but takes ownership of the traversable structure, which can avoid unnecessary clones in certain situations.

Type Parameters

F: The applicative functor containing the effects

Returns

The reordered structure wrapped in a single effect

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.