rustica::transformers

Module cont_t

Continuation Monad Transformer (ContT)

The ContT monad transformer generalizes the continuation monad to operate over an arbitrary monad M. It enables computations in continuation-passing style to be composed with effects from another monad.

Type

ContT<R, M, A> wraps a function (A -> M<R>) -> M<R>

Use Cases

Composing advanced control flow with side effects (IO, state, etc.)

Implementing early exit, backtracking, coroutines within effectful computations

Examples

use rustica::transformers::cont_t::ContT;
use rustica::datatypes::id::Id;

// Create a simple ContT over Id
let cont = ContT::<i32, Id<i32>, i32>::pure(42);
let result = cont.run(|x| Id::new(x * 2));
assert_eq!(result, Id::new(84));

// Chain ContT computations
let cont2 = cont.bind(|x| ContT::pure(x + 1));
let result2 = cont2.run(|x| Id::new(x * 2));
assert_eq!(result2, Id::new(86));

Structs

ContT

The continuation monad transformer: ContT<R, M, A> wraps a function (A -> M<R>) -> M<R>.

Type Aliases

ContTFn

Type alias for the core continuation transformer function type