rustica::utils::hkt_utils

Function fan_out

pub fn fan_out<A: Clone, B, F>(input: A, operations: Vec<F>) -> Vec<B>
where
    F: Fn(A) -> B,

Applies multiple operations to a single input value.

This function takes a value and a collection of operations, applies each operation to the input value, and collects the results. This is useful when you need to transform a single input in multiple different ways.

Type Parameters

A: The type of the input value

B: The type of the output values

F: The type of the operation functions

Arguments

input: The input value to transform

operations: A collection of operations to apply

Returns

A vector containing the results of applying each operation to the input

Examples

use rustica::utils::hkt_utils::fan_out;

// Single input value
let value = 10;

// Multiple operations to apply
let ops = vec![
    |n: i32| n + 5,
    |n: i32| n * 2,
    |n: i32| n * n
];

// Apply all operations to the input
let results = fan_out(value, ops);

// We get the result of each operation
assert_eq!(results, vec![15, 20, 100]);