rustica::utils::hkt_utils

Function zip_with

pub fn zip_with<A, B, C, F>(xs: Vec<A>, ys: Vec<B>, f: F) -> Vec<C>
where
    F: Fn(A, B) -> C,

Combines elements from two collections using a combining function.

This function pairs elements from two collections and applies a combining function to each pair, producing a new collection of the results. The length of the output will be the minimum of the lengths of the two input collections.

Type Parameters

A: The type of elements in the first input collection

B: The type of elements in the second input collection

C: The type of elements in the output collection

F: The type of the combining function

Arguments

xs: The first input collection

ys: The second input collection

f: A function that combines elements from both collections

Returns

A new vector containing the combined elements

Examples

use rustica::utils::hkt_utils::zip_with;

let numbers = vec![1, 2, 3, 4];
let multipliers = vec![10, 20, 30];

// Multiply corresponding elements
let result = zip_with(
    numbers,
    multipliers,
    |n, m| n * m
);

assert_eq!(result, vec![10, 40, 90]);  // Note: only 3 elements (minimum length)