rustica::utils::transform_utils

Function transform_all

pub fn transform_all<T, F, U>(values: &[T], f: F) -> Vec<T::Output<U>>
where
    T: Functor,
    F: Fn(&T::Source) -> U + Copy,
    U: Clone,

Applies a transformation to all functorial values in a collection.

This function maps a transformation over each functorial value in the collection, using the Functor trait’s capabilities.

Type Parameters

T: A type that implements Functor

F: The transformation function type

U: The result type of the transformation

Arguments

values: The collection of functorial values to transform

f: The transformation function to apply to the content of each functor

Returns

A new vector containing the transformed functorial values

Examples

use rustica::utils::transform_utils::transform_all;
use rustica::prelude::*;
use rustica::datatypes::maybe::Maybe;

// Create a collection of Maybe values
let values: Vec<Maybe<i32>> = vec![
    Maybe::Just(1),
    Maybe::Just(2),
    Maybe::Nothing,
    Maybe::Just(4)
];

// Transform the values inside each Maybe
let doubles = transform_all(&values, |&x| x * 2);

// The transformation is applied only to the Just values
assert_eq!(doubles[0], Maybe::Just(2));
assert_eq!(doubles[1], Maybe::Just(4));
assert_eq!(doubles[2], Maybe::Nothing);
assert_eq!(doubles[3], Maybe::Just(8));

Works with standard library types too:

use rustica::utils::transform_utils::transform_all;
use rustica::prelude::*;

// A collection of Option values
let options: Vec<Option<String>> = vec![
    Some("hello".to_string()),
    None,
    Some("world".to_string())
];

// Transform the values inside each Option
let uppercase = transform_all(&options, |s| s.to_uppercase());

assert_eq!(uppercase[0], Some("HELLO".to_string()));
assert_eq!(uppercase[1], None);
assert_eq!(uppercase[2], Some("WORLD".to_string()));