rustica::utils::transform_utils

Function transform_chain

pub fn transform_chain<T, F, U>(value: Option<T>, f: F) -> Option<T::Output<U>>
where
    T: Functor,
    F: Fn(&T::Source) -> U,
    U: Clone,

Applies a transformation to a single optional value.

This function maps a transformation over the value inside an Option functor, preserving the None case and transforming the Some case.

Type Parameters

T: A type that implements Functor

F: The transformation function type

U: The result type of the transformation

Arguments

value: The optional functorial value to transform

f: The transformation function to apply to the content

Returns

An optional transformed functorial value, or None if the input was None

Examples

use rustica::utils::transform_utils::transform_chain;
use rustica::prelude::*;
use rustica::datatypes::maybe::Maybe;

// Apply a transformation to a Just value
let maybe_just: Option<Maybe<i32>> = Some(Maybe::Just(5));
let result1 = transform_chain(maybe_just, |&x| x * 2);
assert_eq!(result1, Some(Maybe::Just(10)));

// Apply a transformation to a Nothing value
let maybe_nothing: Option<Maybe<i32>> = Some(Maybe::Nothing);
let result2 = transform_chain(maybe_nothing, |&x| x * 2);
assert_eq!(result2, Some(Maybe::Nothing));

// When the outer Option is None, the result is None
let none: Option<Maybe<i32>> = None;
let result3 = transform_chain(none, |&x| x * 2);
assert_eq!(result3, None);