rustica::utils::transform_utils

Struct Pipeline

pub struct Pipeline<T>(/* private fields */);

A pipeline for building chains of transformations on functorial values.

This type provides a fluent interface for applying a sequence of transformations to a functorial value. It wraps the value and allows chaining transformations before finally extracting the result.

Type Parameters

T: The type of the functorial value in the pipeline

Examples

use rustica::utils::transform_utils::Pipeline;
use rustica::prelude::*;

// Create a pipeline with an Option value
let pipeline = Pipeline::new(Some(5));

// Apply a series of transformations
let result = pipeline
    .map(|&x| x * 2)      // Double the value
    .map(|x| x.to_string()) // Convert to string
    .extract();

assert_eq!(result, Some("10".to_string()));

// Works with Either as well
use rustica::datatypes::either::Either;

let err_pipeline = Pipeline::new(Either::<&str, i32>::left("error"));
let err_result = err_pipeline
    .map(|&x| x * 2)
    .extract();

assert_eq!(err_result, Either::left("error"));

let ok_pipeline = Pipeline::new(Either::<&str, i32>::right(7));
let ok_result = ok_pipeline
    .map(|&x| x * 2)
    .extract();

assert_eq!(ok_result, Either::right(14));

Implementations

impl<T> Pipeline<T>

pub fn new(value: T) -> Self

Creates a new pipeline with the given functorial value.

Arguments

value: The initial functorial value for the pipeline

Returns

A new pipeline containing the value

Examples

use rustica::utils::transform_utils::Pipeline;
use rustica::datatypes::maybe::Maybe;

// Create a pipeline with a Maybe value
let pipeline: Pipeline<Maybe<i32>> = Pipeline::new(Maybe::Just(42));

// Create a pipeline with an Option value
let option_pipeline: Pipeline<Option<String>> = Pipeline::new(Some("hello".to_string()));

pub fn extract(self) -> T

Extracts the final value from the pipeline.

This method returns the transformed value after all pipeline operations have been applied.

Returns

The final transformed functorial value

Examples

use rustica::utils::transform_utils::Pipeline;
use rustica::prelude::*;

// Create and extract from a pipeline
let result = Pipeline::new(Some(10))
    .extract();

assert_eq!(result, Some(10));

impl<T: Functor> Pipeline<T>

pub fn map_owned<F, U>(self, f: F) -> Pipeline<T::Output<U>>
where F: Fn(T::Source) -> U, U: Clone,

Maps a transformation over the inner value, taking ownership.

This method applies a transformation to the content of the functorial value using the owned version of fmap. This is useful when you donโt need to preserve the original value.

Type Parameters

F: The transformation function type

U: The result type of the transformation

Arguments

f: The transformation function to apply

Returns

A new pipeline containing the transformed value

Examples

use rustica::utils::transform_utils::Pipeline;
use rustica::prelude::*;

// Transform with map_owned
let result = Pipeline::new(Some(5))
    .map_owned(|x| x * 3)  // Using value by value
    .extract();

assert_eq!(result, Some(15));

pub fn map<F, U>(self, f: F) -> Pipeline<T::Output<U>>
where F: Fn(&T::Source) -> U, U: Clone,

Maps a transformation over the inner value, operating by reference.

This method applies a transformation to the content of the functorial value using the reference version of fmap. This preserves the pattern of operating on values by reference.

Type Parameters

F: The transformation function type

U: The result type of the transformation

Arguments

f: The transformation function to apply

Returns

A new pipeline containing the transformed value

Examples

use rustica::utils::transform_utils::Pipeline;
use rustica::prelude::*;
use rustica::datatypes::either::Either;

// Transform with map
let result = Pipeline::new(Either::<String, i32>::right(7))
    .map(|&x| x.to_string())  // Using value by reference
    .extract();

assert_eq!(result, Either::<String, String>::right("7".to_string()));