rustica::datatypes::wrapper::predicate

Struct Predicate

pub struct Predicate<A> { /* private fields */ }

An intensional set, defined by a predicate function.

A Predicate<A> wraps a function that determines whether a value of type A satisfies some condition. Predicates can be combined using logical operations like AND, OR, NOT, and difference.

Implementations

impl<A> Predicate<A>

pub fn new<F>(f: F) -> Self
where F: Fn(&A) -> bool + 'static,

Create a new predicate from a closure.

Parameters

f: A function that evaluates whether a value of type A satisfies the predicate.

Returns

A new Predicate<A> wrapping the provided function.

Example

use rustica::datatypes::wrapper::predicate::Predicate;

let is_even = Predicate::new(|x: &i32| *x % 2 == 0);
assert!(is_even.contains(&2));
assert!(!is_even.contains(&3));

pub fn contains(&self, a: &A) -> bool

Returns true if the value satisfies the predicate.

Arguments

a - The value to test against the predicate

Returns

true if the value satisfies the predicate; false otherwise.

Examples

use rustica::datatypes::wrapper::predicate::Predicate;

let is_positive = Predicate::new(|x: &i32| *x > 0);
assert!(is_positive.contains(&5));
assert!(!is_positive.contains(&-3));

pub fn union(&self, other: &Predicate<A>) -> Predicate<A>
where A: 'static,

Returns a predicate which is the union of this predicate and another.

The union predicate evaluates to true if either this predicate or the other predicate evaluates to true for a given input.

Arguments

other - Another predicate to union with this one

Returns

A new Predicate<A> representing the union of both predicates

Examples

use rustica::datatypes::wrapper::predicate::Predicate;

let is_even = Predicate::new(|x: &i32| *x % 2 == 0);
let is_positive = Predicate::new(|x: &i32| *x > 0);
let even_or_positive = is_even.union(&is_positive);

assert!(even_or_positive.contains(&2));  // Even and positive
assert!(even_or_positive.contains(&-4)); // Even but not positive
assert!(even_or_positive.contains(&3));  // Positive but not even
assert!(!even_or_positive.contains(&-5)); // Neither even nor positive

pub fn intersection(&self, other: &Predicate<A>) -> Predicate<A>
where A: 'static,

Returns a predicate which is the intersection of this predicate and another.

The intersection predicate evaluates to true only if both this predicate and the other predicate evaluate to true for a given input.

Arguments

other - Another predicate to intersect with this one

Returns

A new Predicate<A> representing the intersection of both predicates

Examples

use rustica::datatypes::wrapper::predicate::Predicate;

let is_even = Predicate::new(|x: &i32| *x % 2 == 0);
let is_positive = Predicate::new(|x: &i32| *x > 0);
let even_and_positive = is_even.intersection(&is_positive);

assert!(even_and_positive.contains(&2));  // Even and positive
assert!(!even_and_positive.contains(&-4)); // Even but not positive
assert!(!even_and_positive.contains(&3));  // Positive but not even
assert!(!even_and_positive.contains(&-5)); // Neither even nor positive

pub fn diff(&self, remove: &Predicate<A>) -> Predicate<A>
where A: 'static,

Returns a predicate which is the set difference of this predicate and another.

The resulting predicate evaluates to true for inputs that satisfy this predicate but do not satisfy the remove predicate.

Arguments

remove - The predicate to subtract from this one

Returns

A new Predicate<A> representing the set difference

Examples

use rustica::datatypes::wrapper::predicate::Predicate;

let is_integer = Predicate::new(|x: &f64| x.fract() == 0.0);
let is_negative = Predicate::new(|x: &f64| *x < 0.0);
let positive_integers = is_integer.diff(&is_negative);

assert!(positive_integers.contains(&2.0));  // Integer and not negative
assert!(!positive_integers.contains(&-3.0)); // Integer but negative
assert!(!positive_integers.contains(&1.5));  // Not an integer

pub fn negate(&self) -> Predicate<A>
where A: 'static,

Returns a predicate that is the negation of this predicate.

The resulting predicate evaluates to true for inputs where this predicate evaluates to false, and vice versa.

Returns

A new Predicate<A> representing the logical negation

Examples

use rustica::datatypes::wrapper::predicate::Predicate;

let is_even = Predicate::new(|x: &i32| *x % 2 == 0);
let is_odd = is_even.negate();

assert!(!is_odd.contains(&2));
assert!(is_odd.contains(&3));